refactor: Spike dashboard JS module split: extract common.js + module-l… [ORB-00145]
Spike dashboard JS module split: extract common.js + module-load app.js

Planned-By: claude
