+++ crates/orbit-cli/assets/dashboard/index.html
@@ -2402,6 +2402,6 @@
       </main>
     </section>
     <div id="gantt-tooltip" role="tooltip" aria-hidden="true"></div>
-    <script src="/static/app.js"></script>
+    <script type="module" src="/static/app.js"></script>
   </body>
 </html>
